stm32/rcc: rename common to bus
