+++ embassy-stm32/src/rcc/bus.rs
